clean up Arrow example entrypoints
