+++ examples/arrow/arrow-particles/arrow-particle-source.ts
@@ -1,0 +1,55 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+import {
+  createStreamingParticleRecordBatchIterator,
+  makeArrowParticleRecordBatches,
+  STREAMING_PARTICLE_BATCH_COUNT
+} from './arrow-particle-data';
+import type {ArrowParticleRendererProps} from './arrow-particle-renderer';
+import {ArrowParticlesControlPanel, makeArrowParticlesControlPanelHtml} from './control-panel';
+import {ArrowExamplePanelManager} from '../arrow-example-panels';
+
+/** Owns the particle stream and its table-inspection UI. */
+export class ArrowParticleSource {
+  readonly controlPanel = new ArrowParticlesControlPanel();
+  readonly panels = new ArrowExamplePanelManager({
+    descriptionHtml: makeArrowParticlesControlPanelHtml()
+  });
+  private isFinalized = false;
+
+  constructor(private readonly onSourceChange: (props: ArrowParticleRendererProps) => void) {}
+
+  initialize(): void {
+    this.panels.mount();
+    this.controlPanel.initialize();
+    this.restart();
+  }
+
+  restart(): void {
+    if (this.isFinalized) return;
+    this.controlPanel.setStreamingBatchStatus(0, STREAMING_PARTICLE_BATCH_COUNT);
+    const recordBatches = makeArrowParticleRecordBatches();
+    const tableStream = this.panels.beginLoadedTableStream({
+      id: 'particles-source',
+      label: 'Loaded particle source',
+      kind: 'source',
+      recordBatches
+    });
+    this.onSourceChange({
+      data: createStreamingParticleRecordBatchIterator(recordBatches),
+      onDataBatch: ({loadedBatchCount}) => {
+        if (this.isFinalized) return;
+        tableStream.setLoadedBatchCount(loadedBatchCount);
+        this.controlPanel.setStreamingBatchStatus(loadedBatchCount, STREAMING_PARTICLE_BATCH_COUNT);
+      }
+    });
+  }
+
+  finalize(): void {
+    this.isFinalized = true;
+    this.controlPanel.destroy();
+    this.panels.finalize();
+  }
+}
