+++ examples/arrow/arrow-polygons/app.ts
@@ -5,21 +5,10 @@
 import type {Device} from '@luma.gl/core';
 import type {AnimationProps} from '@luma.gl/engine';
 import {AnimationLoopTemplate} from '@luma.gl/engine';
-import {
-  createStreamingPolygonRecordBatchIterator,
-  type ArrowPolygonColorKind,
-  type ArrowPolygonRowCountKind,
-  type ArrowPolygonSourceKind,
-  type ArrowPolygonViewState,
-  makeArrowPolygonExampleData
-} from './arrow-polygon-data';
-import {
-  ArrowPolygonRenderer,
-  type ArrowPolygonRendererModel,
-  type ArrowPolygonRendererPickingInfo
-} from '@luma.gl/arrow';
-import {ArrowExamplePanelManager, makeArrowExamplePanelHostHtml} from '../arrow-example-panels';
-import {ArrowPolygonControlPanel} from './control-panel';
+import {ArrowPolygonRenderer, type ArrowPolygonRendererPickingInfo} from '@luma.gl/arrow';
+import type {ArrowPolygonViewState} from './arrow-polygon-data';
+import {ArrowPolygonSource} from './arrow-polygon-source';
+import {makeArrowExamplePanelHostHtml} from '../arrow-example-panels';
 
 export const title = 'Polygons';
 export const description =
@@ -27,63 +16,41 @@ export const description =
 
 export default class ArrowPolygonAnimationLoopTemplate extends AnimationLoopTemplate {
   static info = makeArrowExamplePanelHostHtml();
-
   static props = {useDevicePixels: true};
-
-  readonly device: Device;
-  readonly controlPanel: ArrowPolygonControlPanel;
-  readonly panels = new ArrowExamplePanelManager({
-    descriptionPanel: () => this.controlPanel.makeDescriptionPanel(),
-    settingsPanel: () => this.controlPanel.makeSettingsPanel()
-  });
-  layer: ArrowPolygonRenderer;
-  rowCountKind: ArrowPolygonRowCountKind = '10k-stream';
-  sourceKind: ArrowPolygonSourceKind = 'polygon';
-  colorKind: ArrowPolygonColorKind = 'row-colors';
-  modelKind: ArrowPolygonRendererModel = 'attribute';
+  readonly layer: ArrowPolygonRenderer;
+  readonly source: ArrowPolygonSource;
   viewState: ArrowPolygonViewState | null = null;
   animationSeconds = 0;
   lastRenderSeconds: number | null = null;
-  isFinalized = false;
 
   constructor({device}: AnimationProps) {
     super();
-    this.device = device as Device;
-    this.layer = new ArrowPolygonRenderer(this.device, {
-      model: this.modelKind,
+    this.layer = new ArrowPolygonRenderer(device as Device, {
+      model: 'attribute',
       onPick: this.handlePolygonPicked
     });
-    this.controlPanel = new ArrowPolygonControlPanel({
-      device: this.device,
-      initialState: {
-        rowCountKind: this.rowCountKind,
-        sourceKind: this.sourceKind,
-        colorKind: this.colorKind,
-        modelKind: this.modelKind
+    this.source = new ArrowPolygonSource(
+      device as Device,
+      update => {
+        this.viewState = update.viewState;
+        this.animationSeconds = 0;
+        this.lastRenderSeconds = null;
+        this.layer.setProps(update);
       },
-      handlers: {
-        onRowCountKindChange: this.handleRowCountKindChange,
-        onSourceKindChange: this.handleSourceKindChange,
-        onColorKindChange: this.handleColorKindChange,
-        onModelKindChange: this.handleModelKindChange
-      },
-      onRefresh: () => this.panels.refresh()
-    });
+      props => this.layer.setProps(props)
+    );
   }
 
   override async onInitialize(): Promise<void> {
-    this.panels.mount();
-    this.controlPanel.initialize();
-    this.streamPolygonInput(this.rowCountKind, this.sourceKind, this.colorKind);
+    this.source.initialize();
   }
 
   override onRender({aspect, device, time, _mousePosition}: AnimationProps): void {
-    this.updateAnimationTime(time);
-    this.layer.setProps({
-      center: this.getScrollCenter(),
-      scale: this.viewState?.scale ?? 1
-    });
-
+    const seconds = time / 1000;
+    if (this.lastRenderSeconds === null) this.lastRenderSeconds = seconds;
+    this.animationSeconds += Math.max(seconds - this.lastRenderSeconds, 0);
+    this.lastRenderSeconds = seconds;
+    this.layer.setProps({center: this.getScrollCenter(), scale: this.viewState?.scale ?? 1});
     const renderPass = device.beginRenderPass({clearColor: [0.02, 0.04, 0.08, 1]});
     this.layer.draw(renderPass, {aspect});
     renderPass.end();
@@ -91,139 +58,31 @@ export default class ArrowPolygonAnimationLoopTemplate extends AnimationLoopTemp
   }
 
   override onFinalize(): void {
-    this.isFinalized = true;
-    this.controlPanel.destroy();
-    this.panels.finalize();
+    this.source.finalize();
     this.layer.destroy();
   }
 
-  private streamPolygonInput(
-    rowCountKind: ArrowPolygonRowCountKind,
-    sourceKind: ArrowPolygonSourceKind,
-    colorKind: ArrowPolygonColorKind
-  ): void {
-    const effectiveColorKind = getEffectiveColorKind(sourceKind, colorKind);
-    this.rowCountKind = rowCountKind;
-    this.sourceKind = sourceKind;
-    this.colorKind = effectiveColorKind;
-    const sourceData = makeArrowPolygonExampleData(rowCountKind, sourceKind, effectiveColorKind);
-    this.viewState = sourceData.viewState;
-    this.animationSeconds = 0;
-    this.lastRenderSeconds = null;
-    this.layer.setProps({
-      tessellated: sourceData.tessellated,
-      colors: effectiveColorKind === 'constant' ? null : undefined,
-      model: this.modelKind,
-      center: sourceData.viewState.startCenter,
-      scale: sourceData.viewState.scale
-    });
-    this.controlPanel.syncControls({rowCountKind, sourceKind, colorKind: effectiveColorKind});
-    this.controlPanel.setPickedLabel('Hover polygon');
-    this.controlPanel.setStreamingBatchStatus(0, sourceData.batchCount);
-    this.updateMetrics();
-    const polygonTableStream = this.panels.beginLoadedTableStream({
-      id: 'polygons-source',
-      label: 'Loaded polygon source',
-      kind: 'source',
-      recordBatches: sourceData.recordBatches
-    });
-
-    this.layer.setProps({
-      data: createStreamingPolygonRecordBatchIterator(sourceData.recordBatches)[
-        Symbol.asyncIterator
-      ](),
-      onDataBatch: ({loadedBatchCount, metrics}) => {
-        if (this.isFinalized) {
-          return;
-        }
-        polygonTableStream.setLoadedBatchCount(loadedBatchCount);
-        this.controlPanel.setStreamingBatchStatus(loadedBatchCount, sourceData.batchCount);
-        this.controlPanel.setMetrics(metrics);
-      }
-    });
-  }
-
-  private updateMetrics(): void {
-    this.controlPanel.setMetrics(this.layer.getMetrics());
-  }
-
-  private updateAnimationTime(timeMilliseconds: number): void {
-    const seconds = timeMilliseconds / 1000;
-    if (this.lastRenderSeconds === null) {
-      this.lastRenderSeconds = seconds;
-    }
-    const elapsedSeconds = Math.max(seconds - this.lastRenderSeconds, 0);
-    this.lastRenderSeconds = seconds;
-    this.animationSeconds += elapsedSeconds;
-  }
-
   private getScrollCenter(): [number, number] {
-    if (!this.viewState) {
-      return [0, 0];
-    }
-
+    if (!this.viewState) return [0, 0];
     const cycleDurationSeconds = this.viewState.scrollDurationSeconds * 2;
     const cyclePosition =
       cycleDurationSeconds > 0
         ? (this.animationSeconds % cycleDurationSeconds) / cycleDurationSeconds
         : 0;
     const scrollProgress = cyclePosition <= 0.5 ? cyclePosition * 2 : (1 - cyclePosition) * 2;
-    const easedProgress = scrollProgress * scrollProgress * (3 - 2 * scrollProgress);
+    const progress = scrollProgress * scrollProgress * (3 - 2 * scrollProgress);
     return [
-      interpolate(this.viewState.startCenter[0], this.viewState.endCenter[0], easedProgress),
-      interpolate(this.viewState.startCenter[1], this.viewState.endCenter[1], easedProgress)
+      this.viewState.startCenter[0] +
+        (this.viewState.endCenter[0] - this.viewState.startCenter[0]) * progress,
+      this.viewState.startCenter[1] +
+        (this.viewState.endCenter[1] - this.viewState.startCenter[1]) * progress
     ];
   }
-
-  private readonly handleRowCountKindChange = (rowCountKind: ArrowPolygonRowCountKind): void => {
-    if (rowCountKind === this.rowCountKind) {
-      return;
-    }
-    this.streamPolygonInput(rowCountKind, this.sourceKind, this.colorKind);
-  };
-
-  private readonly handleSourceKindChange = (sourceKind: ArrowPolygonSourceKind): void => {
-    if (sourceKind === this.sourceKind) {
-      return;
-    }
-    this.streamPolygonInput(this.rowCountKind, sourceKind, this.colorKind);
-  };
-
-  private readonly handleColorKindChange = (colorKind: ArrowPolygonColorKind): void => {
-    if (colorKind === this.colorKind) {
-      return;
-    }
-    this.streamPolygonInput(this.rowCountKind, this.sourceKind, colorKind);
-  };
-
-  private readonly handleModelKindChange = (modelKind: ArrowPolygonRendererModel): void => {
-    if (modelKind === this.modelKind) {
-      return;
-    }
-    this.modelKind = modelKind;
-    this.controlPanel.syncControls({modelKind});
-    this.layer.setProps({model: modelKind});
-  };
 
   private readonly handlePolygonPicked = ({
     batchIndex,
     rowIndex
   }: ArrowPolygonRendererPickingInfo): void => {
-    this.controlPanel.setPickedLabel(
-      batchIndex === null || rowIndex === null
-        ? 'Hover polygon'
-        : `row ${rowIndex.toLocaleString()} / batch ${(batchIndex + 1).toLocaleString()}`
-    );
+    this.source.setPickedRow(batchIndex, rowIndex);
   };
-}
-
-function interpolate(start: number, end: number, progress: number): number {
-  return start + (end - start) * progress;
-}
-
-function getEffectiveColorKind(
-  sourceKind: ArrowPolygonSourceKind,
-  colorKind: ArrowPolygonColorKind
-): ArrowPolygonColorKind {
-  return sourceKind === 'dense-union' && colorKind === 'vertex-colors' ? 'row-colors' : colorKind;
 }
