+++ examples/deck/arrow-path-layer/app.ts
@@ -5,23 +5,11 @@
 import {Deck, OrthographicView} from '@deck.gl/core';
 import {ArrowPathLayer} from '@deck.gl-community/arrow-layers';
 import {getArrowLayerTooltip} from '../arrow-layer-tooltip';
-import {
-  makeArrowLineRecordBatches,
-  makeArrowLineSourceData,
-  PATH_DATASETS
-} from '../../arrow/arrow-lines/arrow-line-data';
+import {initializeArrowPathLayerSource} from './arrow-path-layer-source';
 
 /** Creates the standalone or website-hosted Deck path-layer example. */
 export function createArrowPathLayerDeck(parent?: HTMLDivElement) {
-  const sourceData = makeArrowLineSourceData(
-    PATH_DATASETS['240'],
-    'lines',
-    'float32',
-    'row-colors',
-    'none'
-  );
-
-  return new Deck({
+  const deck = new Deck({
     parent,
     views: new OrthographicView({id: 'main'}),
     initialViewState: {target: [0, 0], zoom: 8},
@@ -31,9 +19,15 @@ export function createArrowPathLayerDeck(parent?: HTMLDivElement) {
       new ArrowPathLayer({
         id: 'arrow-paths',
         pickable: true,
-        data: makeArrowLineRecordBatches(sourceData),
+        data: [],
         model: 'attribute'
       })
     ]
   });
+  initializeArrowPathLayerSource(data => {
+    deck.setProps({
+      layers: [new ArrowPathLayer({id: 'arrow-paths', pickable: true, data, model: 'attribute'})]
+    });
+  });
+  return deck;
 }
